refactor(wp): move getRelatedArticles to Post
