+++ db/wpdb.ts
@@ -16,7 +16,6 @@ import { registerExitHandler } from "./cleanup.js"
 import {
     RelatedChart,
     WP_PostType,
-    PostReference,
     JsonError,
     PostRestApi,
     uniqBy,
@@ -402,72 +401,6 @@ export const getRelatedResearchAndWritingForVariable = async (
     // uses different charts that all use a single indicator we would get duplicates for the post to link to so
     // here we deduplicate by url. The first item is retained by uniqBy, latter ones are discarded.
     return uniqBy(allSortedRelatedResearch, "url")
-}
-
-/*
- * Get all the related research and writing for a chart
- */
-export const getRelatedArticles = async (
-    chartId: number
-): Promise<PostReference[] | undefined> => {
-    const relatedPosts: PostReference[] = (
-        await db.knexInstance().raw(
-            `
-        SELECT
-            p.title,
-            p.slug,
-            p.id
-        FROM
-            posts_with_gdoc_publish_status p
-            JOIN posts_links pl ON p.id = pl.sourceId
-            JOIN charts c ON pl.target = c.slug
-            OR pl.target IN (
-                SELECT
-                    cr.slug
-                FROM
-                    chart_slug_redirects cr
-                WHERE
-                    cr.chart_id = c.id
-            )
-        WHERE
-            c.id = ?
-            AND p.status = 'publish'
-            AND p.isGdocPublished = 0
-            AND p.type != 'wp_block'
-            -- note: we are not filtering by linkType to cast of wider net: if a post links to an
-            -- explorer having the same slug as the grapher chart, we want to surface it as
-            -- a "Related research" as it is most likely relevant.
-        UNION
-        SELECT
-            pg.content ->> '$.title' AS title,
-            pg.slug AS slug,
-            pg.id AS id
-        FROM
-            posts_gdocs pg
-            JOIN posts_gdocs_links pgl ON pg.id = pgl.sourceId
-            JOIN charts c ON pgl.target = c.slug
-            OR pgl.target IN (
-                SELECT
-                    cr.slug
-                FROM
-                    chart_slug_redirects cr
-                WHERE
-                    cr.chart_id = c.id
-            )
-        WHERE
-            c.id = ?
-            AND pg.content ->> '$.type' <> 'fragment'
-            AND pg.published = 1
-            -- note: we are not filtering by linkType here either, for the same reason as above.
-        `,
-            [chartId, chartId]
-        )
-    )[0]
-
-    return uniqBy(relatedPosts, "slug").sort(
-        // Alphabetise
-        (a, b) => (a.title.toLowerCase() > b.title.toLowerCase() ? 1 : -1)
-    )
 }
 
 export const getBlockApiFromApi = async (id: number): Promise<any> => {
